Auto-import the keys
